Dev release

+++ routes/web.php
@@ -14,8 +14,8 @@ Route::get('/home', 'HomeController@index')->name('home');
 Route::get('/profile', 'ProfileController@index')->name('profile');
 Route::put('/profile', 'ProfileController@update')->name('profile.update');
 
-
-Route::group(['prefix' => 'categories', 'middleware' => 'auth'], function() {
+// Categories Routes
+Route::group(['prefix' => 'categories', 'middleware' => 'auth', 'middleware' => 'role:Admin'], function() {
     Route::get('/', 'CategoryController@index')->name('categories.index');
     Route::get('/create', 'CategoryController@create')->name('categories.create');
     Route::post('/', 'CategoryController@store')->name('categories.store');
@@ -24,16 +24,18 @@ Route::group(['prefix' => 'categories', 'middleware' => 'auth'], function() {
     Route::delete('/{id}', 'CategoryController@destroy')->name('categories.destroy');
 });
 
+// Products Routes
 Route::group(['prefix' => 'products', 'middleware' => 'auth'], function() {
     Route::get('/', 'ProductController@index')->name('products.index');
-    Route::get('/create', 'ProductController@create')->name('products.create');
-    Route::post('/', 'ProductController@store')->name('products.store');
+    Route::get('/create', 'ProductController@create')->name('products.create')->middleware('permission:Add Products');
+    Route::post('/', 'ProductController@store')->name('products.store')->middleware('permission:Add Products');
     Route::get('/{id}/edit', 'ProductController@edit')->name('products.edit');
     Route::put('/{id}', 'ProductController@update')->name('products.update');
-    Route::delete('/{id}', 'ProductController@destroy')->name('products.destroy');
+    Route::delete('/{id}', 'ProductController@destroy')->name('products.destroy')->middleware('permission:Delete Products');
 });
 
-Route::group(['prefix' => 'transactions', 'middleware' => 'auth'], function() {
+// Transactions Routes
+Route::group(['prefix' => 'transactions', 'middleware' => 'auth', 'middleware' => 'permission:Transaction'], function() {
     Route::get('/', 'TransactionController@index')->name('transactions.index');
     Route::get('/create', 'TransactionController@create')->name('transactions.create');
     Route::post('/', 'TransactionController@store')->name('transactions.store');
@@ -41,10 +43,40 @@ Route::group(['prefix' => 'transactions', 'middleware' => 'auth'], function() {
     Route::get('/{id}/print', 'TransactionController@print')->name('transactions.print');
 });
 
-Route::group(['prefix' => 'reports', 'middleware' => 'auth'], function() {
+// Reports Routes
+Route::group(['prefix' => 'reports', 'middleware' => 'auth', 'middleware' => 'permission:Report Access'], function() {
     Route::get('/monthly', 'ReportController@monthlyReport')->name('reports.monthly');
     Route::get('/products', 'ReportController@productReport')->name('reports.product');
 });
+
+// Roles Routes
+Route::group(['prefix' => 'roles', 'middleware' => 'auth', 'middleware' => 'role:Admin'], function() {
+    Route::get('/', 'RoleController@index')->name('roles.index');
+    Route::get('/create', 'RoleController@create')->name('roles.create');
+    Route::post('/', 'RoleController@store')->name('roles.store');
+    Route::get('/{id}/edit', 'RoleController@edit')->name('roles.edit');
+    Route::put('/{id}', 'RoleController@update')->name('roles.update');
+    Route::delete('/{id}', 'RoleController@destroy')->name('roles.destroy');
+});
+
+// Permissions Routes
+Route::group(['prefix' => 'permissions', 'middleware' => 'auth', 'middleware' => 'role:Admin'], function() {
+    Route::get('/', 'PermissionController@index')->name('permissions.index');
+    Route::get('/create', 'PermissionController@create')->name('permissions.create');
+    Route::post('/', 'PermissionController@store')->name('permissions.store');
+    Route::get('/{id}/edit', 'PermissionController@edit')->name('permissions.edit');
+    Route::put('/{id}', 'PermissionController@update')->name('permissions.update');
+    Route::delete('/{id}', 'PermissionController@destroy')->name('permissions.destroy');
+});
+
+// Users Routes
+Route::group(['prefix' => 'users', 'middleware' => 'auth', 'middleware' => 'role:Admin|Leader'], function() {
+    Route::get('/', 'UserController@index')->name('users.index');
+    Route::get('/{id}/edit', 'UserController@edit')->name('users.edit');
+    Route::put('/{id}', 'UserController@update')->name('users.update');
+    Route::delete('/{id}', 'UserController@destroy')->name('users.destroy');
+});
+
 
 Route::get('/about', function () {
     return view('about');
